+++ TarponSimulator2017/Core/EmptyInterface.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TarponSimulator2017
+{
+	public interface EmptyInterface
+	{
+	}
+}
+
