+++ TarponSimulator2017/Core/Fish.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Tarpon.Core
+{
+	public class Fish : IUpdatable
+	{
+		public Fish ()
+		{
+		}
+	}
+}
+
