+++ TarponSimulator2017/Core/FishFactory.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Tarpon.Core
+{
+	public class FishFactory
+	{
+		public FishFactory ()
+		{
+		}
+	}
+}
+
