+++ TarponSimulator2017/Core/FishingRod.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Tarpon.Core
+{
+	public class FishingRod
+	{
+		public FishingRod ()
+		{
+		}
+	}
+}
+
