+++ TarponSimulator2017/Core/Physics.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Tarpon.Core
+{
+	public class Physics
+	{
+		public Physics ()
+		{
+		}
+	}
+}
+
