+++ TarponSimulator2017/Core/World.cs
@@ -4,27 +4,17 @@ namespace Tarpon.Core
 {
 	public class World
 	{
-		public List<Boat> Boats { get; private set; }
-		private List<IUpdatable> toUpdate;
+		public Boat playerBoat { get; private set; }
+		private List<IUpdatable> toUpdate { get; private set; }
 
 		public World()
 		{
-			Boats = new List<Boat>();
-			toUpdate = new List<IUpdatable>();
+
 		}
 
-		public Boat AddBoat()
+		public void UpdateUpdatables()
 		{
-			Boat b = new Boat();
-			Boats.Add(b);
-			toUpdate.Add(b);
-
-			return b;
-		}
-
-		public void Update(int now)
-		{
-			toUpdate.ForEach(tu => tu.Update(now));
+			
 		}
 	}
 }
