+++ TarponSimulator2017/Core/WorldBuilder.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Tarpon.Core
+{
+	public class WorldBuilder
+	{
+		
+	}
+}
