+++ TarponSimulator2017/EmptyClass.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace TarponSimulator2017
+{
+	public class EmptyClass
+	{
+		public EmptyClass ()
+		{
+		}
+	}
+}
+
