+++ TarponSimulator2017/TarponSimulator2017.csproj
@@ -8,6 +8,8 @@
     <RootNamespace>TarponSimulator2017</RootNamespace>
     <AssemblyName>TarponSimulator2017</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <ProductVersion>8.0.30703</ProductVersion>
+    <SchemaVersion>2.0</SchemaVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
     <DebugSymbols>true</DebugSymbols>
@@ -59,10 +61,14 @@
     <Compile Include="Controller\WorldController.cs" />
     <Compile Include="Draw\IDrawer.cs" />
     <Compile Include="Draw\Utils\PictureLoader.cs" />
-    <Compile Include="Core\Elements\Boat.cs" />
-    <Compile Include="Core\Builders\WorldBuilder.cs" />
     <Compile Include="Core\World.cs" />
     <Compile Include="Core\IUpdatable.cs" />
+    <Compile Include="Core\Boat.cs" />
+    <Compile Include="Core\WorldBuilder.cs" />
+    <Compile Include="Core\Fish.cs" />
+    <Compile Include="Core\FishFactory.cs" />
+    <Compile Include="Core\FishingRod.cs" />
+    <Compile Include="Core\Physics.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Content\" />
@@ -70,7 +76,5 @@
     <Folder Include="Draw\" />
     <Folder Include="Controller\" />
     <Folder Include="Draw\Utils\" />
-    <Folder Include="Core\Builders\" />
-    <Folder Include="Core\Elements\" />
   </ItemGroup>
 </Project>
